pyproject.toml
[project]
name = "backend2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "clerk-backend-api==3.1.9",
    "cryptography>=44.0.1,<45.0.0",
    "dj-database-url>=3.0.1",
    "django>=5.2.8",
    "django-activity-stream>=2.0.0",
    "django-cors-headers>=4.9.0",
    "django-environ>=0.12.0",
    "djangorestframework>=3.16.1",
    "psycopg2-binary>=2.9.11",
    "pyjwt>=2.10.1",
    "pytz>=2025.2",
    "requests>=2.32.5",
    "svix>=1.81.0",
]
